Fix : 호텔페이지이동
