+++ frontend/src/Component/Header/index.css
@@ -11,7 +11,7 @@
   z-index: 999;
 }
 .header-content:hover{
-  background-color: rgba(255, 255, 255, 0.9);
+  background-color: rgba(226, 226, 226, 0.9);
 }
 .header-content.scrolled {
   background-image: linear-gradient(to bottom, rgba(255, 255, 255, 0.9), rgba(255, 255, 255, 0));
