+++ frontend/src/Page/TravelPackagePage/product/productdetail/index.css
@@ -1,6 +1,145 @@
-.productdetailpage{
-    margin: 80px 0 0 80px;
+.productdetail-page{
+    width: 100%;
 }
-.product-img{
-    width: 300px;
+.productdetail-page_product-img{
+    width: 100%;
+    border-radius: 25px;
+}
+.productdetail-page_detail-img{
+    width: 600px;
+}
+.productdetail-page_contnet{
+    display: flex;
+    justify-content: center;
+    padding: 20px 30px;
+    margin: 0 30px;
+    border-radius: 20px;
+    box-shadow: 0px 1px 3px 0px #333;
+    align-items: center;
+}
+.productdetail-page_contnet_review-content h3{
+    text-align: center;
+}
+.productdetail-page_contnet_review {
+    padding: 10px;
+    height: 400px;
+    overflow: auto;
+}
+.productdetail-page_contnet_review ul {
+    list-style-type: none;
+    padding: 20px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+.productdetail-page_contnet_review li {
+    position: relative;
+    width: 80%;
+    font-size: 1.15rem;
+    font-weight: 700;
+    margin-bottom: 25px;
+    padding: 15px;
+    background-color: rgb(90, 120, 250);
+    border-radius: 8px;
+    color: white;
+    font-family: "TiltNeon-Regular-VariableFont_XROT,YROT";
+}
+.productdetail-page_contnet_review li::after {
+    content:"";
+    border-top: 15px solid rgb(90, 120, 250);
+    border-left: 15px solid transparent;
+    border-right: 0px solid transparent;
+    border-bottom: 0px solid transparent;
+    position:absolute;
+    top : 10px;
+    left: -15px;
+}
+
+/*메인페이지 2번쩨*/
+.productdetail-page2{
+    display: flex;
+    justify-content: space-around;
+    margin: 80px 60px;
+    align-items: flex-start;
+}
+.reservation-form {
+    position: sticky;
+    top: 74px;
+    padding:  20px 40px;
+    box-shadow: 0px 1px 3px 0px #333;
+    border-radius: 10px;
+    text-align: center;
+}
+
+
+  /* 테이블 스타일 */
+.reservation-form table {
+    margin: 10px 0;
+}
+.reservation-form td,
+.reservation-form th {
+    padding: 4px 0;
+    text-align: right;
+    border: none;
+}
+.reservation-form td:nth-child(1){
+    text-align: left;
+}
+.reservation-form-input{
+    padding: 0 25px;
+}
+.reservation-form select {
+    margin: 0;
+    padding: 5px 0;
+    border: none;
+    text-align: center;
+    width: 100%;
+    border: 1px solid #333;
+}
+.counter-button{
+    margin: 8px;
+    padding: 0 6.5px;
+    font-size: 0.75rem;
+    text-align: center;
+    border: none;
+    border-radius: 10px;
+    color: white;
+    background-color: rgb(90, 120, 250);
+    cursor: pointer;
+}
+.reservation-btn{
+    width: 100%;
+    border-radius: 10px;
+    padding: 5px 0;
+    margin-bottom: 10px;
+    color: #f0f0f0;
+    background-color: rgb(90, 120, 250);
+    border: none;
+}
+.reservation-form button:hover {
+    background-color: #0056b3;
+}
+.petcount_add{
+    font-size: 1.05rem;
+    font-weight: 700;
+}
+.petsize-info{
+    text-align: left;
+    font-size: 0.75rem;
+    margin: 20px 0;
+}
+.petsize-info div{
+    margin-top: -10px;
+}
+.petsize-info div p{
+    margin: 0;
+}
+.petsize-info button{
+    background:none;
+    border: none;
+}
+.petsize-info button:hover{
+    color: blue;
+    background: none;
+    text-decoration: underline;
 }
